added type price comparison, adfuller needs work
A comparison between the organic and the conventional type avocado has been added, with a short explaination of the process. The Adfuller test is not working correctly and needs further work, perhaps more later on. It is important to evaluate the stationarty of the data before the model is put in production.